+++ src/App/WalletSuite/ConnectedWallet/Details/WalletDetails/Balances/CoinBalances.tsx
