+++ src/App/WalletSuite/ConnectedWallet/Details/WalletDetails/Balances/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Balances";
